requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.2
WTForms==3.1.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
email-validator==2.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
waitress==2.1.2
